+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/Calendar.cs
@@ -2,8 +2,8 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class Calendar : AuditEntity<long>
 {
-    public int UserId { get; set; }
-    public int? AddEventsFromTeamId { get; set; }
+    public long UserId { get; set; }
+    public long? AddEventsFromTeamId { get; set; }
     public bool CheckForConflicts { get; set; }
     
     public User User { get; set; }
