+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/ExternalAttendee.cs
@@ -2,7 +2,7 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class ExternalAttendee : Entity<long>
 {
-    public int AvailabilitySlotId { get; set; }
+    public long AvailabilitySlotId { get; set; }
     public DateTimeOffset EventTime { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
