+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/ExternalAttendeeAvailability.cs
@@ -2,7 +2,7 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class ExternalAttendeeAvailability : Entity<long>
 {
-    public int ExternalAttendeeId { get; set; }
+    public long ExternalAttendeeId { get; set; }
     public DateTimeOffset StartEvent { get; set; }
     public DateTimeOffset EndEvent { get; set; }
