+++ backend/EasyMeets.Core/EasyMeets.Core.DAL/Entities/TeamMemberMeeting.cs
@@ -2,8 +2,8 @@ namespace EasyMeets.Core.DAL.Entities;
 
 public class TeamMemberMeeting : Entity<long>
 {
-    public int MemberId { get; set; }
-    public int EventId { get; set; }
+    public long MemberId { get; set; }
+    public long EventId { get; set; }
     public int Priority { get; set; }
     
     public User User { get; }
